Fixed nil InternalContainerLifecycle in cm stubs.
