+++ pkg/kubelet/cm/container_manager_stub.go
@@ -21,6 +21,7 @@ import (
 	"k8s.io/api/core/v1"
 
 	internalapi "k8s.io/kubernetes/pkg/kubelet/apis/cri"
+	"k8s.io/kubernetes/pkg/kubelet/cm/cpumanager"
 	kubecontainer "k8s.io/kubernetes/pkg/kubelet/container"
 	"k8s.io/kubernetes/pkg/kubelet/status"
 )
@@ -75,7 +76,7 @@ func (cm *containerManagerStub) GetResources(pod *v1.Pod, container *v1.Containe
 }
 
 func (cm *containerManagerStub) InternalContainerLifecycle() InternalContainerLifecycle {
-	return nil
+	return &internalContainerLifecycleImpl{cpumanager.NewFakeManager()}
 }
 
 func NewStubContainerManager() ContainerManager {
