+++ pkg/kubelet/cm/container_manager_unsupported.go
@@ -25,6 +25,7 @@ import (
 	"k8s.io/client-go/tools/record"
 	internalapi "k8s.io/kubernetes/pkg/kubelet/apis/cri"
 	"k8s.io/kubernetes/pkg/kubelet/cadvisor"
+	"k8s.io/kubernetes/pkg/kubelet/cm/cpumanager"
 	kubecontainer "k8s.io/kubernetes/pkg/kubelet/container"
 	"k8s.io/kubernetes/pkg/kubelet/status"
 	"k8s.io/kubernetes/pkg/util/mount"
@@ -80,7 +81,7 @@ func (cm *unsupportedContainerManager) GetResources(pod *v1.Pod, container *v1.C
 }
 
 func (cm *unsupportedContainerManager) InternalContainerLifecycle() InternalContainerLifecycle {
-	return nil
+	return &internalContainerLifecycleImpl{cpumanager.NewFakeManager()}
 }
 
 func NewContainerManager(_ mount.Interface, _ cadvisor.Interface, _ NodeConfig, failSwapOn bool, devicePluginEnabled bool, recorder record.EventRecorder) (ContainerManager, error) {
